# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PPA Valuation Engine API",
    "description": "Dutch renewable energy PPA valuation and asset optimisation",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/solver/negotiation-range": {
      "post": {
        "tags": [
          "solver"
        ],
        "summary": "Negotiation Range",
        "operationId": "negotiation_range_api_solver_negotiation_range_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NegotiationRangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NegotiationRangeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/solver/tornado": {
      "post": {
        "tags": [
          "solver"
        ],
        "summary": "Tornado",
        "operationId": "tornado_api_solver_tornado_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TornadoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TornadoResultSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/solver/fair-strike": {
      "post": {
        "tags": [
          "solver"
        ],
        "summary": "Fair Strike",
        "operationId": "fair_strike_api_solver_fair_strike_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FairStrikeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FairStrikeResultSchema"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DealConfigSchema": {
        "properties": {
          "start_date": {
            "type": "string",
            "title": "Start Date",
            "default": "2027-01-01"
          },
          "end_date": {
            "type": "string",
            "title": "End Date",
            "default": "2036-12-31"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "EUR"
          },
          "discount_rate": {
            "type": "number",
            "exclusiveMaximum": 0.3,
            "exclusiveMinimum": 0.0,
            "title": "Discount Rate",
            "default": 0.06
          }
        },
        "type": "object",
        "title": "DealConfigSchema"
      },
      "FairStrikeRequest": {
        "properties": {
          "config": {
            "$ref": "#/components/schemas/PPAConfigSchema",
            "default": {
              "location": {
                "altitude": 10.0,
                "lat": 52.1,
                "lon": 5.1,
                "tz": "Europe/Amsterdam"
              },
              "solar": {
                "azimuth_deg": 180.0,
                "capacity_mw": 50.0,
                "cloud_factor_logit_scale": 0.7,
                "cloud_factor_phi": 0.9,
                "degradation_rate": 0.005,
                "monthly_cloud_means": [
                  0.38,
                  0.42,
                  0.5,
                  0.57,
                  0.63,
                  0.68,
                  0.68,
                  0.65,
                  0.58,
                  0.5,
                  0.42,
                  0.36
                ],
                "performance_ratio": 0.8,
                "seed": 42,
                "tilt_deg": 35.0
              },
              "market": {
                "ar1_phi": 0.7,
                "ar1_sigma": 8.0,
                "base_price": 80.0,
                "cannibalisation_alpha_2027": 55.0,
                "cannibalisation_alpha_2036": 90.0,
                "drift": -1.5,
                "seasonal_amplitude": 15.0,
                "seasonal_peak_day": 15,
                "seed": 42,
                "weekday_hourly_adder": [
                  -10.0,
                  -12.0,
                  -14.0,
                  -14.0,
                  -12.0,
                  -10.0,
                  -8.0,
                  8.0,
                  20.0,
                  14.0,
                  8.0,
                  2.0,
                  -5.0,
                  -5.0,
                  0.0,
                  8.0,
                  15.0,
                  35.0,
                  30.0,
                  22.0,
                  14.0,
                  5.0,
                  0.0,
                  -8.0
                ],
                "weekend_hourly_adder": [
                  -12.0,
                  -14.0,
                  -16.0,
                  -16.0,
                  -14.0,
                  -12.0,
                  -10.0,
                  -5.0,
                  2.0,
                  5.0,
                  8.0,
                  8.0,
                  2.0,
                  2.0,
                  2.0,
                  4.0,
                  6.0,
                  10.0,
                  12.0,
                  10.0,
                  6.0,
                  2.0,
                  -2.0,
                  -10.0
                ]
              },
              "load": {
                "annual_consumption_gwh": 100.0,
                "noise_sigma": 0.05,
                "seasonal_amplitude": 0.15,
                "seasonal_peak_day": 15,
                "seed": 42,
                "weekday_hourly_shape": [
                  0.55,
                  0.5,
                  0.48,
                  0.48,
                  0.52,
                  0.62,
                  0.78,
                  0.9,
                  0.98,
                  1.02,
                  1.1,
                  1.06,
                  1.0,
                  1.02,
                  1.0,
                  1.02,
                  1.05,
                  1.08,
                  1.1,
                  1.04,
                  0.92,
                  0.8,
                  0.7,
                  0.6
                ],
                "weekend_hourly_shape": [
                  0.55,
                  0.5,
                  0.48,
                  0.48,
                  0.5,
                  0.58,
                  0.68,
                  0.78,
                  0.85,
                  0.9,
                  0.92,
                  0.9,
                  0.88,
                  0.85,
                  0.83,
                  0.83,
                  0.85,
                  0.88,
                  0.88,
                  0.83,
                  0.75,
                  0.68,
                  0.62,
                  0.58
                ]
              },
              "deal": {
                "currency": "EUR",
                "discount_rate": 0.06,
                "end_date": "2036-12-31",
                "start_date": "2027-01-01"
              }
            }
          },
          "structure_a": {
            "type": "string",
            "enum": [
              "pay_as_produced",
              "pay_as_nominated",
              "baseload"
            ],
            "title": "Structure A",
            "default": "pay_as_produced"
          },
          "structure_b": {
            "type": "string",
            "enum": [
              "pay_as_produced",
              "pay_as_nominated",
              "baseload"
            ],
            "title": "Structure B",
            "default": "baseload"
          },
          "pricing_structure": {
            "type": "string",
            "enum": [
              "fixed_flat",
              "fixed_escalated"
            ],
            "title": "Pricing Structure",
            "default": "fixed_flat"
          },
          "strike_a": {
            "type": "number",
            "title": "Strike A",
            "default": 65.0
          },
          "n_paths": {
            "type": "integer",
            "title": "N Paths",
            "default": 200
          }
        },
        "type": "object",
        "title": "FairStrikeRequest"
      },
      "FairStrikeResultSchema": {
        "properties": {
          "structure_a": {
            "type": "string",
            "title": "Structure A"
          },
          "structure_b": {
            "type": "string",
            "title": "Structure B"
          },
          "strike_a": {
            "type": "number",
            "title": "Strike A"
          },
          "fair_strike_b": {
            "type": "number",
            "title": "Fair Strike B"
          },
          "p50_npv_a": {
            "type": "number",
            "title": "P50 Npv A"
          },
          "p50_npv_b": {
            "type": "number",
            "title": "P50 Npv B"
          },
          "pricing_structure": {
            "type": "string",
            "title": "Pricing Structure"
          },
          "n_paths": {
            "type": "integer",
            "title": "N Paths"
          }
        },
        "type": "object",
        "required": [
          "structure_a",
          "structure_b",
          "strike_a",
          "fair_strike_b",
          "p50_npv_a",
          "p50_npv_b",
          "pricing_structure",
          "n_paths"
        ],
        "title": "FairStrikeResultSchema"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LocationConfigSchema": {
        "properties": {
          "lat": {
            "type": "number",
            "title": "Lat",
            "default": 52.1
          },
          "lon": {
            "type": "number",
            "title": "Lon",
            "default": 5.1
          },
          "tz": {
            "type": "string",
            "title": "Tz",
            "default": "Europe/Amsterdam"
          },
          "altitude": {
            "type": "number",
            "title": "Altitude",
            "default": 10.0
          }
        },
        "type": "object",
        "title": "LocationConfigSchema"
      },
...
      "NegotiationRangeRequest": {
        "properties": {
          "config": {
            "$ref": "#/components/schemas/PPAConfigSchema",
            "default": {
              "location": {
                "altitude": 10.0,
                "lat": 52.1,
                "lon": 5.1,
                "tz": "Europe/Amsterdam"
              },
              "solar": {
                "azimuth_deg": 180.0,
                "capacity_mw": 50.0,
                "cloud_factor_logit_scale": 0.7,
                "cloud_factor_phi": 0.9,
                "degradation_rate": 0.005,
                "monthly_cloud_means": [
                  0.38,
                  0.42,
                  0.5,
                  0.57,
                  0.63,
                  0.68,
                  0.68,
                  0.65,
                  0.58,
                  0.5,
                  0.42,
                  0.36
                ],
                "performance_ratio": 0.8,
                "seed": 42,
                "tilt_deg": 35.0
              },
              "market": {
                "ar1_phi": 0.7,
                "ar1_sigma": 8.0,
                "base_price": 80.0,
                "cannibalisation_alpha_2027": 55.0,
                "cannibalisation_alpha_2036": 90.0,
                "drift": -1.5,
                "seasonal_amplitude": 15.0,
                "seasonal_peak_day": 15,
                "seed": 42,
                "weekday_hourly_adder": [
                  -10.0,
                  -12.0,
                  -14.0,
                  -14.0,
                  -12.0,
                  -10.0,
                  -8.0,
                  8.0,
                  20.0,
                  14.0,
                  8.0,
                  2.0,
                  -5.0,
                  -5.0,
                  0.0,
                  8.0,
                  15.0,
                  35.0,
                  30.0,
                  22.0,
                  14.0,
                  5.0,
                  0.0,
                  -8.0
                ],
                "weekend_hourly_adder": [
                  -12.0,
                  -14.0,
                  -16.0,
                  -16.0,
                  -14.0,
                  -12.0,
                  -10.0,
                  -5.0,
                  2.0,
                  5.0,
                  8.0,
                  8.0,
                  2.0,
                  2.0,
                  2.0,
                  4.0,
                  6.0,
                  10.0,
                  12.0,
                  10.0,
                  6.0,
                  2.0,
                  -2.0,
                  -10.0
                ]
              },
              "load": {
                "annual_consumption_gwh": 100.0,
                "noise_sigma": 0.05,
                "seasonal_amplitude": 0.15,
                "seasonal_peak_day": 15,
                "seed": 42,
                "weekday_hourly_shape": [
                  0.55,
                  0.5,
                  0.48,
                  0.48,
                  0.52,
                  0.62,
                  0.78,
                  0.9,
                  0.98,
                  1.02,
                  1.1,
                  1.06,
                  1.0,
                  1.02,
                  1.0,
                  1.02,
                  1.05,
                  1.08,
                  1.1,
                  1.04,
                  0.92,
                  0.8,
                  0.7,
                  0.6
                ],
                "weekend_hourly_shape": [
                  0.55,
                  0.5,
                  0.48,
                  0.48,
                  0.5,
                  0.58,
                  0.68,
                  0.78,
                  0.85,
                  0.9,
                  0.92,
                  0.9,
                  0.88,
                  0.85,
                  0.83,
                  0.83,
                  0.85,
                  0.88,
                  0.88,
                  0.83,
                  0.75,
                  0.68,
                  0.62,
                  0.58
                ]
              },
              "deal": {
                "currency": "EUR",
                "discount_rate": 0.06,
                "end_date": "2036-12-31",
                "start_date": "2027-01-01"
              }
            }
          },
          "target_producer_npv": {
            "type": "number",
            "title": "Target Producer Npv",
            "default": 0.0
          }
        },
        "type": "object",
        "title": "NegotiationRangeRequest"
      },
      "NegotiationRangeResponse": {
        "properties": {
          "ranges": {
            "items": {
              "$ref": "#/components/schemas/NegotiationRangeSchema"
            },
            "type": "array",
            "title": "Ranges"
          }
        },
        "type": "object",
        "required": [
          "ranges"
        ],
        "title": "NegotiationRangeResponse"
      },
      "NegotiationRangeSchema": {
        "properties": {
          "supply_structure": {
            "type": "string",
            "title": "Supply Structure"
          },
          "producer_floor": {
            "$ref": "#/components/schemas/SolverResultSchema"
          },
          "consumer_ceiling": {
            "$ref": "#/components/schemas/SolverResultSchema"
          },
          "midpoint": {
            "type": "number",
            "title": "Midpoint"
          },
          "spread_eur_mwh": {
            "type": "number",
            "title": "Spread Eur Mwh"
          }
        },
        "type": "object",
        "required": [
          "supply_structure",
          "producer_floor",
          "consumer_ceiling",
          "midpoint",
          "spread_eur_mwh"
        ],
        "title": "NegotiationRangeSchema"
      },
      "PPAConfigSchema": {
        "properties": {
          "location": {
            "$ref": "#/components/schemas/LocationConfigSchema"
          },
          "solar": {
            "$ref": "#/components/schemas/SolarConfigSchema"
          },
          "market": {
            "$ref": "#/components/schemas/MarketPriceConfigSchema"
          },
          "load": {
            "$ref": "#/components/schemas/ConsumerLoadConfigSchema"
          },
          "deal": {
            "$ref": "#/components/schemas/DealConfigSchema"
          }
        },
        "type": "object",
        "title": "PPAConfigSchema"
      },
...
      "SolarConfigSchema": {
        "properties": {
          "capacity_mw": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Capacity Mw",
            "default": 50.0
          },
          "tilt_deg": {
            "type": "number",
            "title": "Tilt Deg",
            "default": 35.0
          },
          "azimuth_deg": {
            "type": "number",
            "title": "Azimuth Deg",
            "default": 180.0
          },
          "performance_ratio": {
            "type": "number",
            "exclusiveMaximum": 1.0,
            "exclusiveMinimum": 0.0,
            "title": "Performance Ratio",
            "default": 0.8
          },
          "degradation_rate": {
            "type": "number",
            "exclusiveMaximum": 0.02,
            "exclusiveMinimum": 0.0,
            "title": "Degradation Rate",
            "default": 0.005
          },
          "cloud_factor_phi": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Cloud Factor Phi",
            "default": 0.9
          },
          "cloud_factor_logit_scale": {
            "type": "number",
            "title": "Cloud Factor Logit Scale",
            "default": 0.7
          },
          "monthly_cloud_means": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "maxItems": 12,
            "minItems": 12,
            "title": "Monthly Cloud Means",
            "default": [
              0.38,
              0.42,
              0.5,
              0.57,
              0.63,
              0.68,
              0.68,
              0.65,
              0.58,
              0.5,
              0.42,
              0.36
            ]
          },
          "seed": {
            "type": "integer",
            "title": "Seed",
            "default": 42
          }
        },
        "type": "object",
        "title": "SolarConfigSchema"
      },
      "SolverResultSchema": {
        "properties": {
          "strike_eur_mwh": {
            "type": "number",
            "title": "Strike Eur Mwh"
          },
          "achieved_npv": {
            "type": "number",
            "title": "Achieved Npv"
          },
          "target_npv": {
            "type": "number",
            "title": "Target Npv"
          },
          "n_iterations": {
            "type": "integer",
            "title": "N Iterations"
          },
          "converged": {
            "type": "boolean",
            "title": "Converged"
          },
          "supply_structure": {
            "type": "string",
            "title": "Supply Structure"
          },
          "pricing_structure": {
            "type": "string",
            "title": "Pricing Structure"
          }
        },
        "type": "object",
        "required": [
          "strike_eur_mwh",
          "achieved_npv",
          "target_npv",
          "n_iterations",
          "converged",
          "supply_structure",
          "pricing_structure"
        ],
        "title": "SolverResultSchema"
      },
...
      "TornadoEntrySchema": {
        "properties": {
          "parameter": {
            "type": "string",
            "title": "Parameter"
          },
          "base_value": {
            "type": "number",
            "title": "Base Value"
          },
          "low_value": {
            "type": "number",
            "title": "Low Value"
          },
          "high_value": {
            "type": "number",
            "title": "High Value"
          },
          "npv_base": {
            "type": "number",
            "title": "Npv Base"
          },
          "npv_low": {
            "type": "number",
            "title": "Npv Low"
          },
          "npv_high": {
            "type": "number",
            "title": "Npv High"
          },
          "delta_npv_low": {
            "type": "number",
            "title": "Delta Npv Low"
          },
          "delta_npv_high": {
            "type": "number",
            "title": "Delta Npv High"
          },
          "abs_swing": {
            "type": "number",
            "title": "Abs Swing"
          }
        },
        "type": "object",
        "required": [
          "parameter",
          "base_value",
          "low_value",
          "high_value",
          "npv_base",
          "npv_low",
          "npv_high",
          "delta_npv_low",
          "delta_npv_high",
          "abs_swing"
        ],
        "title": "TornadoEntrySchema"
      },
      "TornadoRequest": {
        "properties": {
          "config": {
            "$ref": "#/components/schemas/PPAConfigSchema",
            "default": {
              "location": {
                "altitude": 10.0,
                "lat": 52.1,
                "lon": 5.1,
                "tz": "Europe/Amsterdam"
              },
              "solar": {
                "azimuth_deg": 180.0,
                "capacity_mw": 50.0,
                "cloud_factor_logit_scale": 0.7,
                "cloud_factor_phi": 0.9,
                "degradation_rate": 0.005,
                "monthly_cloud_means": [
                  0.38,
                  0.42,
                  0.5,
                  0.57,
                  0.63,
                  0.68,
                  0.68,
                  0.65,
                  0.58,
                  0.5,
                  0.42,
                  0.36
                ],
                "performance_ratio": 0.8,
                "seed": 42,
                "tilt_deg": 35.0
              },
              "market": {
                "ar1_phi": 0.7,
                "ar1_sigma": 8.0,
                "base_price": 80.0,
                "cannibalisation_alpha_2027": 55.0,
                "cannibalisation_alpha_2036": 90.0,
                "drift": -1.5,
                "seasonal_amplitude": 15.0,
                "seasonal_peak_day": 15,
                "seed": 42,
                "weekday_hourly_adder": [
                  -10.0,
                  -12.0,
                  -14.0,
                  -14.0,
                  -12.0,
                  -10.0,
                  -8.0,
                  8.0,
                  20.0,
                  14.0,
                  8.0,
                  2.0,
                  -5.0,
                  -5.0,
                  0.0,
                  8.0,
                  15.0,
                  35.0,
                  30.0,
                  22.0,
                  14.0,
                  5.0,
                  0.0,
                  -8.0
                ],
                "weekend_hourly_adder": [
                  -12.0,
                  -14.0,
                  -16.0,
                  -16.0,
                  -14.0,
                  -12.0,
                  -10.0,
                  -5.0,
                  2.0,
                  5.0,
                  8.0,
                  8.0,
                  2.0,
                  2.0,
                  2.0,
                  4.0,
                  6.0,
                  10.0,
                  12.0,
                  10.0,
                  6.0,
                  2.0,
                  -2.0,
                  -10.0
                ]
              },
              "load": {
                "annual_consumption_gwh": 100.0,
                "noise_sigma": 0.05,
                "seasonal_amplitude": 0.15,
                "seasonal_peak_day": 15,
                "seed": 42,
                "weekday_hourly_shape": [
                  0.55,
                  0.5,
                  0.48,
                  0.48,
                  0.52,
                  0.62,
                  0.78,
                  0.9,
                  0.98,
                  1.02,
                  1.1,
                  1.06,
                  1.0,
                  1.02,
                  1.0,
                  1.02,
                  1.05,
                  1.08,
                  1.1,
                  1.04,
                  0.92,
                  0.8,
                  0.7,
                  0.6
                ],
                "weekend_hourly_shape": [
                  0.55,
                  0.5,
                  0.48,
                  0.48,
                  0.5,
                  0.58,
                  0.68,
                  0.78,
                  0.85,
                  0.9,
                  0.92,
                  0.9,
                  0.88,
                  0.85,
                  0.83,
                  0.83,
                  0.85,
                  0.88,
                  0.88,
                  0.83,
                  0.75,
                  0.68,
                  0.62,
                  0.58
                ]
              },
              "deal": {
                "currency": "EUR",
                "discount_rate": 0.06,
                "end_date": "2036-12-31",
                "start_date": "2027-01-01"
              }
            }
          },
          "supply_structure": {
            "type": "string",
            "enum": [
              "pay_as_produced",
              "pay_as_nominated",
              "baseload"
            ],
            "title": "Supply Structure",
            "default": "pay_as_produced"
          },
          "pricing_structure": {
            "type": "string",
            "enum": [
              "fixed_flat",
              "fixed_escalated",
              "indexed_with_floor_cap",
              "floating"
            ],
            "title": "Pricing Structure",
            "default": "fixed_flat"
          },
          "base_strike": {
            "type": "number",
            "title": "Base Strike",
            "default": 65.0
          }
        },
        "type": "object",
        "title": "TornadoRequest"
      },
      "TornadoResultSchema": {
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/TornadoEntrySchema"
            },
            "type": "array",
            "title": "Entries"
          },
          "supply_structure": {
            "type": "string",
            "title": "Supply Structure"
          },
          "pricing_structure": {
            "type": "string",
            "title": "Pricing Structure"
          },
          "base_strike": {
            "type": "number",
            "title": "Base Strike"
          }
        },
        "type": "object",
        "required": [
          "entries",
          "supply_structure",
          "pricing_structure",
          "base_strike"
        ],
        "title": "TornadoResultSchema"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}